requirements.txt
hurry.filesize
numpy
paramiko
